api/requirements.txt
astroid==2.9.3
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gitdb==4.0.9
GitPython==3.1.26
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
platformdirs==2.4.1
pylint==2.12.2
smmap==5.0.0
toml==0.10.2
typing-extensions==4.0.1
Werkzeug==2.0.2
wrapt==1.13.3
